if logged in user can only modify or delete his account
sticky forms on form and login

+++ Exercice1Revisions_Marcelo/public_html/UserFunctions.php
@@ -19,22 +19,14 @@ function ShowFormModif($user) {
     echo '<br><label for = "prenom">Prénom:</label><input id = "prenom" type = "text" placeholder = "Prenom" name = "prenom" value="' . $prenom . '" required/><br>';
     echo '<br><label for = "date">Date de naissance:</label><input id = "date" type = "date" name = "date" value="' . $date . '" required/><br>';
     echo '<br><label for = "pseudo">Pseudo:</label><input id = "pseudo" type = "text" placeholder = "Pseudo" name = "pseudo" value="' . $pseudo . '" required/><br>';
-    echo '<br><label for = "mdp">Password:</label><input id = "mdp" type = "password" placeholder = "Modifie pas le mdp" name = "mdp"  value="' . $mdp . '" disabled/><br>';
+    echo '<br><label for = "mdp">Password:</label><input id = "mdp" type = "password" placeholder = "Modifie pas le mdp" name = "mdp"  value="" disabled/><br>';
     echo '<br><label for = "email">Email:</label><input id = "email" type = "email" placeholder = "Email" name = "email" value="' . $email . '" required/><br>';
     echo '<br><label for = "description">Description:</label><textarea id = "description" rows = "5" cols = "25" placeholder = "description" name = "description">' . $description . '</textarea><br>';
     echo '<input type = "submit" name = "modifButton" value = "Modify"/>';
     echo '<input type = "button" name = "annuler" value = "Annuler"/><br/>';
-    echo '<a href = "users.php">Utilisateurs</a><br />';    
+    echo '<a href = "users.php">Utilisateurs</a><br />';
 }
 
-function ShowFormConnection() {
-    echo '<br><label for = "pseudo">Pseudo:</label><input id = "pseudo" type = "text" placeholder = "Pseudo" name = "pseudo" required/><br>';
-    echo '<br><label for = "mdp">Password:</label><input id = "mdp" type = "password" placeholder = "Mot de passe" name = "mdp" required/> <br><br>';
-    echo '<input type = "submit" name = "connecter" value = "Se connecter"/>';
-    echo '<input type = "button" name = "annuler" value = "Annuler"/><br><br>';
-    echo '<a href = "users.php">Utilisateurs</a><br />  ';
-    echo '<a href = "index.php">Formulaire</a>';
-}
 function ShowUser() {
     $dataAll = getAllFields();
 
@@ -44,13 +36,15 @@ function ShowUser() {
         echo '<td>' . $user['Prenom'] . '</td>';
         echo '<td>' . $user['Email'] . '</td>';
         echo '<td><a href="users.php?id=' . $user['ID'] . '">Details</a></td>';
-        echo '<td><a href="index.php?id=' . $user['ID'] . '" name="modifLink">Modifier</a></td>';
-        echo '<td><a href="users.php?idDelete=' . $user['ID'] . '" name="deleteUser">Supprimer</a></td>';
+        if (isset($_SESSION['userlogin']) && $_SESSION['userlogin'] == $user['Pseudo']) {
+            echo '<td><a href="index.php?id=' . $user['ID'] . '" name="modifLink">Modifier</a></td>';
+            echo '<td><a href="users.php?idDelete=' . $user['ID'] . '" name="deleteUser">Supprimer</a></td>';
+        }
         echo '</tr>';
     }
 }
 
-function ShowUserDetails($user) {    
+function ShowUserDetails($user) {
 
     echo '<tr>';
     echo '<td>' . $user['Nom'] . '</td>';
